delete finished, w/o error msg
