+++ WebApp/package.json
@@ -34,17 +34,14 @@
     "@material-ui/core": "^3.2.0",
     "@material-ui/icons": "^3.0.1",
     "axios": "^0.18.0",
-    "bootstrap": "^4.3.1",
     "classnames": "^2.2.5",
     "downshift": "^3.2.10",
-    "jquery": "^1.9.1",
     "lodash": "^4.17.5",
     "material-color-hash": "^0.1.4",
     "mlcm": "0.0.3",
     "object-assign": "4.1.0",
-    "react": "^16.8.0",
-    "react-bootstrap": "^1.0.0-beta.14",
-    "react-dom": "^16.8.0",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
     "react-modal": "^3.11.1",
     "react-onclickoutside": "^6.7.1",
     "react-redux": "5.0.5",
