+++ WebApp/src/components/Toolbar/DeleteAccount/DeleteAccountModal.scss
@@ -1,0 +1,21 @@
+.modal-window {
+  border-radius: 2px;
+  border-color: 1px solid #e0e0e0;
+  color: #000;
+  background-color: #fff;
+  display: "flex";
+  align-items: center;
+  justify-content: center;
+  background: border-box;
+}
+
+.paper {
+  background-color: #fff;
+  padding: 25px;
+  width: fit-content;
+  outline: none;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
