+++ WebApp/src/index.js
@@ -5,7 +5,6 @@ import { render } from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
 import configureStore, { history } from './store/configureStore';
 import Root from './components/Root';
-import 'bootstrap/dist/css/bootstrap.min.css';
 import './styles/styles.scss';
 require('./favicon.ico');
